pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page SEO-afforability analyzer web app"
authors = ["Yaroslav Pekatoros <y.pekatoros@gmail.com>"]
readme = "README.md"
repository = "https://github.com/ya-pekatoros/python-project-83"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
validators = "^0.20.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
